requirements.txt
aioredis
python-dotenv
loguru
vk_api
https://github.com/timoniq/vkbottle/archive/master.zip
redis